wip optimconfig
