+++ src/levanter/optim/__init__.py
@@ -1,0 +1,7 @@
+from .config import AdamConfig, OptimizerConfig
+from .sophia import (  # SophiaGConfig,; SophiaGObjective,
+    ScaleBySophiaState,
+    SophiaHConfig,
+    scale_by_sophia_g,
+    scale_by_sophia_h,
+)
